refactor: renaming and moving EmailDTO to Email
