+++ backend/agricultura-familiar/src/main/java/com/es/agriculturafamiliar/listener/DummyListener.java
@@ -3,7 +3,7 @@ package com.es.agriculturafamiliar.listener;
 import java.util.Map;
 
 import com.es.agriculturafamiliar.constants.TemplateType;
-import com.es.agriculturafamiliar.dto.EmailDTO;
+import com.es.agriculturafamiliar.entity.Email;
 import com.es.agriculturafamiliar.event.DummyEvent;
 import com.es.agriculturafamiliar.service.AsyncMessageService;
 import com.es.agriculturafamiliar.service.EmailMessageService;
@@ -17,7 +17,7 @@ import lombok.AllArgsConstructor;
 @AllArgsConstructor
 @Component
 public class DummyListener {
-    private final AsyncMessageService<EmailDTO> asyncMessageService;
+    private final AsyncMessageService<Email> asyncMessageService;
 
     @EventListener
     @Async
@@ -26,7 +26,7 @@ public class DummyListener {
         String to = dummyEvent.getToEmail();
 
         Map<String, Object> context = Map.of("name", name);
-        EmailDTO emailDTO = EmailDTO.builder()
+        Email email = Email.builder()
             .subject("testEmail")
             .variables(context)
             .to(to)
@@ -34,7 +34,7 @@ public class DummyListener {
             .templateName(TemplateType.REGISTRATION_PRODUTOR)
             .build(); 
                      
-            asyncMessageService.sendMessage(emailDTO);
+            asyncMessageService.sendMessage(email);
             
     }
