+++ backend/agricultura-familiar/src/main/java/com/es/agriculturafamiliar/service/EmailMessageService.java
@@ -3,7 +3,7 @@ package com.es.agriculturafamiliar.service;
 import javax.mail.internet.MimeMessage;
 
 import com.es.agriculturafamiliar.config.AsyncMessageSenderFactory;
-import com.es.agriculturafamiliar.dto.EmailDTO;
+import com.es.agriculturafamiliar.entity.Email;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.mail.javamail.JavaMailSender;
@@ -11,20 +11,20 @@ import org.springframework.scheduling.annotation.Async;
 import org.springframework.stereotype.Service;
 
 @Service
-public class EmailMessageService implements AsyncMessageService<EmailDTO> {
+public class EmailMessageService implements AsyncMessageService<Email> {
 
-    private AsyncMessageSenderFactory<EmailDTO, JavaMailSender, MimeMessage> asyncMessageSenderFactory;
+    private AsyncMessageSenderFactory<Email, JavaMailSender, MimeMessage> asyncMessageSenderFactory;
     private JavaMailSender mailSender;
 
     @Autowired
-    public EmailMessageService(AsyncMessageSenderFactory<EmailDTO, JavaMailSender, MimeMessage> asyncMessageSenderFactory, JavaMailSender mailSender) {        
+    public EmailMessageService(AsyncMessageSenderFactory<Email, JavaMailSender, MimeMessage> asyncMessageSenderFactory, JavaMailSender mailSender) {        
         this.asyncMessageSenderFactory = asyncMessageSenderFactory;
         this.mailSender = mailSender;
     }
     
     @Async
     @Override
-    public void sendMessage(EmailDTO data) {
+    public void sendMessage(Email data) {
         MimeMessage constructedMessage = asyncMessageSenderFactory.createAsyncMessageSender(data, mailSender);
         mailSender.send(constructedMessage);
     }
